refactor(api): inline pipelineTestFixtures into TestDecisionsDataManager
createInstanceWithSchema was redundant: createDecisionSetup already accepts
a processSchema, createInstanceDataFromTemplate already copies selectionPipeline
from template phases, and createDecisionInstance already seeds currentStateId
from the first phase. executeTestTransition moves onto the manager as
advancePhase; createAuthenticatedCaller moves next to its session helpers.
