+++ services/api/src/routers/decision/instances/transitionPipelineProposals.test.ts
@@ -1,6 +1,7 @@
 import { getProposalsForPhase } from '@op/common';
 import { db, eq } from '@op/db/client';
 import {
+  ProcessStatus,
   ProposalStatus,
   decisionTransitionProposals,
   stateTransitionHistory,
@@ -9,12 +10,10 @@ import { describe, expect, it } from 'vitest';
 
 import { TestDecisionsDataManager } from '../../../test/helpers/TestDecisionsDataManager';
 import {
-  createInstanceWithSchema,
-  executeTestTransition,
   schemaWithPipeline,
   schemaWithThreePhasesAndPipelines,
   schemaWithoutPipeline,
-} from '../../../test/helpers/pipelineTestFixtures';
+} from '../../../test/helpers/pipelineSchemas';
 
 describe.concurrent('Transition pipeline: join table population', () => {
   it('creates exactly 2 join rows when selectionPipeline limits to 2 from 3 proposals', async ({
@@ -23,11 +22,13 @@ describe.concurrent('Transition pipeline: join table population', () => {
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
 
-    const { instanceId, userEmail } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
 
     // Create and submit 3 proposals (submitted proposals are eligible for transition)
     for (let i = 1; i <= 3; i++) {
@@ -39,7 +40,7 @@ describe.concurrent('Transition pipeline: join table population', () => {
       });
     }
 
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'submission',
       toPhaseId: 'review',
@@ -70,11 +71,13 @@ describe.concurrent('Transition pipeline: join table population', () => {
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
 
-    const { instanceId, userEmail } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithThreePhasesAndPipelines,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithThreePhasesAndPipelines,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
 
     // Submit 4 proposals; first pipeline limits to 3, second limits to 2
     for (let i = 1; i <= 4; i++) {
@@ -87,14 +90,14 @@ describe.concurrent('Transition pipeline: join table population', () => {
     }
 
     // First transition: submission → review (limit 3)
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'submission',
       toPhaseId: 'review',
     });
 
     // Second transition: review → final (limit 2, from the 3 that survived)
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'review',
       toPhaseId: 'final',
@@ -131,11 +134,13 @@ describe.concurrent('Transition pipeline: join table population', () => {
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
 
-    const { instanceId, userEmail } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithoutPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithoutPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
 
     // Create and submit 3 proposals (submitted proposals are eligible for transition)
     for (let i = 1; i <= 3; i++) {
@@ -147,7 +152,7 @@ describe.concurrent('Transition pipeline: join table population', () => {
       });
     }
 
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'submission',
       toPhaseId: 'review',
@@ -194,11 +199,13 @@ describe.concurrent('Transition pipeline: join table population', () => {
       ],
     };
 
-    const { instanceId, userEmail } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithZeroLimit,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithZeroLimit,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
 
     await testData.createProposal({
       userEmail,
@@ -210,7 +217,7 @@ describe.concurrent('Transition pipeline: join table population', () => {
     // Transition succeeds but pipeline eliminates all proposals.
     // The transition row exists, so this is unambiguously "new system, zero survivors"
     // (not legacy data) — getProposalsForPhase must return [].
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'submission',
       toPhaseId: 'review',
